Refuse a fractional lease, and give the refusal its own code
Two rulings the ports asked for as reference.

THE CODE IS `task_lease_invalid`. Both ports chose that spelling
independently and it is the better one: the refusal previously reused
`unsafe_state_configuration`, which collapses two failures with two different
fixes — change a number, or change a driver — into one code a consumer
cannot branch on. Decision 0020 in a smaller frame. The volatile-store
refusal keeps the old code; only the lease moves, into its own exception.

A FRACTIONAL LEASE IS REFUSED, NOT TRUNCATED. This is the same rule as
refusing a lease of zero, one scale down: `90.4` becoming `90` is a
configuration silently becoming a different configuration, and that is what
was rejected the first time. Truncation landing in the safe direction — a
shorter lease, so more reclaims rather than lost work — is the clamping
argument restated, and it was not enough for zero. A fractional lease could
not have been honoured as written in any case, because `claimed_until` is an
integer Unix timestamp in all three languages.

PHP's own signature already refused a float passed to `claim()`, but not the
path that matters: `(int) '90.4'` is `90`, so the config route truncated
silently. The `(int)` cast is therefore gone from the shipped config too —
otherwise the guard would have been defeated by the file that declares the
setting, which is a test of its own.

`lock_wait` is deliberately left tolerant: a local wait bound that never
reaches a stored record and is not part of the shared contract.

Fixture loops now count their iterations and assert the count. A test-id
listing cannot see inside one test, so a loop that stopped running would
report green having probed nothing.

+++ config/prism-harness.php
@@ -118,7 +118,19 @@ return [
     */
 
     'tasks' => [
-        'lease_seconds' => (int) env('HARNESS_TASK_LEASE_SECONDS', AgentTaskSource::DEFAULT_LEASE_SECONDS),
+        /*
+         * NOT cast to an int here, deliberately.
+         *
+         * `(int) '90.4'` is 90, and the cast would do that silently before
+         * anything had a chance to object — so the guard against a fractional
+         * lease would be defeated by the config file that declares it. The raw
+         * value is passed through and refused where it is read, which is the
+         * only place that can tell 300 from '300' from '90.4'.
+         */
+        'lease_seconds' => env('HARNESS_TASK_LEASE_SECONDS', AgentTaskSource::DEFAULT_LEASE_SECONDS),
+
+        // Cast, because this one is tolerant by decision — see
+        // PrismHarness::taskLockWait().
         'lock_wait' => (int) env('HARNESS_TASK_LOCK_WAIT', 5),
     ],
